v4.0.0 - Bootloader issue solved - get_info displays less information (needed for the hand) - timer_value (now debug) is in microseconds - encoder and control loop are now working at 1KHz - analog input is made at once by usgin a for loop spending less time

+++ firmware.cydsn/globals.c
@@ -33,6 +33,6 @@ float tau_feedback;
 
 // utility timer value for display
 
-uint16 timer_value;
+uint32 timer_value;
 
 /* END OF FILE */
